+++ flake.nix
@@ -2,46 +2,52 @@
   description = "Higher-order, low-boilerplate free monads.";
 
   inputs = {
-    stable.url = github:nixos/nixpkgs/nixos-20.09;
+    nixpkgs_2009.url = github:nixos/nixpkgs/release-20.09;
+    nixpkgs_2105.url = github:nixos/nixpkgs/release-21.05;
     unstable.url = github:nixos/nixpkgs/nixpkgs-unstable;
     flake-utils.url = github:numtide/flake-utils;
   };
 
-  outputs = { stable, unstable, flake-utils, ... }:
+  outputs = { nixpkgs_2009, nixpkgs_2105, unstable, flake-utils, ... }:
   flake-utils.lib.eachSystem ["x86_64-linux"] (system:
+  with unstable.lib;
   let
     hsPkgs = nixpkgs: compiler: import ./nix/overlay.nix { inherit system nixpkgs compiler; };
 
-    ghc865 = hsPkgs stable "ghc865";
-    ghc884 = hsPkgs unstable "ghc884";
-    ghc8104 = hsPkgs unstable "ghc8104";
-    ghc901 = hsPkgs unstable "ghc901";
-
-    packages = {
-      inherit (ghc8104) polysemy polysemy-plugin;
-      polysemy-865 = ghc865.polysemy;
-      polysemy-plugin-865 = ghc865.polysemy-plugin;
-      polysemy-884 = ghc884.polysemy;
-      polysemy-plugin-884 = ghc884.polysemy-plugin;
-      polysemy-8104 = ghc8104.polysemy;
-      polysemy-plugin-8104 = ghc8104.polysemy-plugin;
-      polysemy-901 = ghc901.polysemy;
-      polysemy-plugin-901 = ghc901.polysemy-plugin;
+    ghcs = {
+      "865" = hsPkgs nixpkgs_2009 "ghc865";
+      "884" = hsPkgs nixpkgs_2105 "ghc884";
+      "8107" = hsPkgs unstable "ghc8107";
+      "902" = hsPkgs unstable "ghc902";
+      "924" = hsPkgs unstable "ghc924";
     };
+
+    mkPackages = version: {
+      "polysemy-${version}" = ghcs.${version}.polysemy;
+      "polysemy-plugin-${version}" = ghcs.${version}.polysemy-plugin;
+    };
+
+    defaultPackages = {
+      inherit (ghcs."902") polysemy polysemy-plugin;
+      default = ghcs."902".polysemy;
+    };
+
+    packages = foldl' (l: r: l // r) defaultPackages (map mkPackages (attrNames ghcs));
+
+    mkDevShell = extra: ghc: ghc.shellFor {
+      packages = p: [p.polysemy p.polysemy-plugin];
+      buildInputs = with ghc; [
+        cabal-install
+      ] ++ (if extra then [ghcid haskell-language-server] else []);
+      withHoogle = extra;
+    };
+
+    devShells = mapAttrs' (n: g: nameValuePair "ghc${n}" (mkDevShell (n != "924") g)) ghcs;
+
   in {
     inherit packages;
 
-    defaultPackage = ghc8104.polysemy;
-
-    devShell = ghc8104.shellFor {
-      packages = _: [];
-      buildInputs = with ghc8104; [
-        cabal-install
-        haskell-language-server
-        ghcid
-      ];
-      withHoogle = true;
-    };
+    devShells = devShells // { default = devShells.ghc902; };
 
     checks = packages;
   });