+++ src/Polysemy/Membership.hs
@@ -9,11 +9,9 @@ module Polysemy.Membership
   -- * Using membership
   , subsumeUsing
   , interceptUsing
-  , interceptUsing
   , interceptUsingH
   ) where
 
 import Polysemy.Internal
-import Polysemy.Internal.Combinators
 import Polysemy.Internal.Interpretation
 import Polysemy.Internal.Union